warehouse/metrics-service/pyproject.toml
[project]
name = "metrics-service"
version = "0.1.0"
description = "Metrics calculation service"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "OSO Team", email = "opensource-observer@googlegroups.com" }]
dependencies = [
    "arrow<2.0.0,>=1.3.0",
    "boltons<25.0.0,>=24.0.0",
    "click<9.0.0,>=8.1.7",
    "clickhouse-connect<1.0.0,>=0.7.16",
    "cloud-sql-python-connector[pg8000]<2.0.0,>=1.6.0",
    "dagster<2.0.0,>=1.8.6",
    "dagster-dbt<1.0.0,>=0.24.0",
    "dagster-embedded-elt<1.0.0,>=0.24.0",
    "dagster-gcp<1.0.0,>=0.24.0",
    "dagster-polars<1.0.0,>=0.24.0",
    "dagster-postgres<1.0.0,>=0.24.0",
    "dagster-webserver<2.0.0,>=1.7.16",
    "dask[distributed]<2025.5.0,>=2024.4.2",
    "dask-kubernetes<2025.5.0,>=2024.4.2",
    "dbt-bigquery<2.0.0,>=1.7.0",
    "dbt-core<2.0.0,>=1.8.2",
    "discord-webhook<2.0.0,>=1.3.1",
    "dlt<2.0.0,>=1.3.0",
    "duckdb<2.0.0,>=1.1.0",
    "gcsfs<2025.0.0,>=2024.6.1",
    "google-api-python-client<3.0.0,>=2.116.0",
    "google-cloud-bigquery<4.0.0,>=3.17.1",
    "google-cloud-bigquery-datatransfer<4.0.0,>=3.15.3",
    "google-cloud-resource-manager<2.0.0,>=1.12.3",
    "google-cloud-secret-manager<3.0.0,>=2.20.0",
    "google-cloud-service-management<2.0.0,>=1.8.3",
    "google-cloud-service-usage<2.0.0,>=1.10.3",
    "google-cloud-storage<3.0.0,>=2.14.0",
    "lz4<5.0.0,>=4.3.3",
    "oso-core",
    "oss-directory==0.2.5",
    "pendulum<4.0.0,>=3.0.0",
    "polars==1.24.0",
    "pytest<9.0.0,>=8.2.1",
    "python-dotenv<2.0.0,>=1.0.1",
    "requests<3.0.0,>=2.31.0",
    "ruamel-yaml<1.0.0,>=0.18.6",
    "sqlalchemy<3.0.0,>=2.0.25",
    "textual<1.0.0,>=0.52.1",
    "redis<6.0.0,>=5.0.7",
    "githubkit<1.0.0,>=0.12.4",
    "sqlmesh[trino]<1.0.0,>=0.141.1",
    "dagster-duckdb<1.0.0,>=0.24.0",
    "dagster-duckdb-polars<1.0.0,>=0.24.0",
    "google-cloud-bigquery-storage<3.0.0,>=2.25.0",
    "dagster-sqlmesh==0.16.0",
    "google-auth<3.0.0,>=2.34.0",
    "pillow<11.0.0,>=10.4.0",
    "dagster-k8s<1.0.0,>=0.24.6",
    "pyiceberg[hive,s3fs]<1.0.0,>=0.8.1",
    "connectorx<1.0.0,>=0.4.0",
    "bokeh<4.0.0,>=3.6.1",
    "fastapi[standard]<1.0.0,>=0.115.6",
    "pyee<13.0.0,>=12.1.1",
    "aiotrino<1.0.0,>=0.2.3",
    "pytest-asyncio<1.0.0,>=0.24.0",
    "isort<6.0.0,>=5.13.2",
    "uvicorn[standard]<1.0.0,>=0.32.1",
    "websockets<15.0,>=14.1",
    "pydantic-settings<3.0.0,>=2.7.0",
    "openrank-sdk<1.0.0,>=0.4.0",
    "pandas<3.0.0,>=2.2.3",
    "docker<8.0.0,>=7.1.0",
    "gitpython<4.0.0,>=3.1.44",
    "minio<8.0.0,>=7.2.15",
    "colorlog<7.0.0,>=6.9.0",
    "pyarrow-stubs<18.0,>=17.16",
    "gcloud-aio-storage<10.0.0,>=9.3.0",
    "kr8s==0.20.7",
]

[tool.uv.sources]
oso-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
